Extract reconciliation, reporting, budgets, scheduling, investments
Batches the remaining five non-core modules in one commit. After this
PR only core remains in LegacyMixin; the next (and final) refactor PR
renames it to CoreMixin.

- book/reconciliation.py: set_reconcile_state, get_unreconciled_splits,
  reconcile_account, void_transaction, unvoid_transaction (+ void slot
  storage for unvoid restoration).
- book/reporting.py: spending_by_category, income_by_source,
  balance_sheet, net_worth (point-in-time + time series), cash_flow,
  debt_payoff_plan (avalanche simulation + YETI multiplier), plus the
  _get_account_depth / _get_account_at_depth hierarchy helpers.
- book/budgets.py: list/get/create/set_amount/report/delete, plus
  budget-specific period machinery (_period_to_date_range,
  _current_period, _resolve_periods with q1-q4 / "ytd" / "all").
- book/scheduling.py: scheduled transaction templates (Recurrence
  rows + splits-json slot), list/get_upcoming/create_from/update/
  delete, and FREQUENCY_TO_RECURRENCE / RECURRENCE_TO_FREQUENCY maps.
  create_transaction_from_scheduled calls self.create_transaction
  (resolved via MRO to LegacyMixin, which still holds core).
- book/investments.py: commodities, prices, and lots. Includes the
  full lot lifecycle (create, assign_split_to_lot with auto-close,
  calculate_lot_gain with latest-price fallback, close_lot manual
  cleanup).

Supporting changes:
- book/_base.py: moved _find_commodity into BaseGnuCashBook since
  create_account (core) uses it alongside the full commodity surface
  in InvestmentsMixin — it's a genuinely shared finder.
- book/__init__.py: _MIXIN_MAP now lists all seven extracted modules.
- server.py: dropped the 34 tool wrappers for these modules (906
  lines from 1801) and the associated section banners.
- tools/<module>.py: five new register(mcp, get_book) files, each
  attaching its tools on demand.
- tests/test_tools.py: module-scoped autouse fixture force-loads all
  extracted tool modules and binds each registered tool's .fn back
  onto the server module namespace (test_tools calls them directly
  as server_module.<name>). Teardown removes the registry entries
  added by the fixture so test_modules.py's lazy-loading assertions
  see a clean state regardless of test order.

At-import tool count: 15 (core only). All 60 non-core tools are
lazy-loaded via importlib on first enable.

MRO: GnuCashBook -> LegacyMixin -> AdminMixin -> BudgetsMixin ->
BusinessMixin -> InvestmentsMixin -> ReconciliationMixin ->
ReportingMixin -> SchedulingMixin -> BaseGnuCashBook.

Full suite: 602 passed. Pure code motion — no behavior change.
Live testing skipped per the pattern established in the admin and
business extractions.
