+++ src/gnucash_mcp/server.py
@@ -353,176 +353,7 @@ def list_accounts(
     return result
 
 
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def list_commodities(verbose: bool = False) -> str:
-    """List all commodities (currencies, stocks, etc.) in the book.
-
-    Returns a compact one-line-per-commodity format by default.
-    Use verbose=true for full JSON with fraction, latest prices, etc.
-
-    Args:
-        verbose: If true, return full JSON details for each commodity.
-    """
-    book = get_book()
-    result = book.list_commodities(compact=not verbose)
-    if verbose:
-        return _json(result)
-    return result
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="create", entity_type="commodity")
-def create_commodity(
-    mnemonic: str,
-    fullname: str,
-    namespace: str = "FUND",
-    fraction: int = 10000,
-    cusip: str | None = None,
-) -> str:
-    """Create a new commodity (stock, mutual fund, etc.) in the book.
-
-    Args:
-        mnemonic: Symbol (e.g., "VTSAX", "AAPL"). Must be unique within namespace.
-        fullname: Full name (e.g., "Vanguard Total Stock Market Index Fund").
-        namespace: Grouping category. Common values:
-            - "FUND" for mutual funds (default)
-            - "NASDAQ", "NYSE", "AMEX" for stocks
-            - Any custom string for other assets
-        fraction: Smallest fractional unit. Use:
-            - 1 for whole units only
-            - 100 for 2 decimal places
-            - 10000 for 4 decimal places (default, standard for shares)
-            - 1000000 for 6 decimal places (crypto)
-        cusip: Optional CUSIP/ISIN identifier for the security.
-    """
-    book = get_book()
-    result = book.create_commodity(
-        mnemonic=mnemonic,
-        fullname=fullname,
-        namespace=namespace,
-        fraction=fraction,
-        cusip=cusip,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="create", entity_type="price")
-def create_price(
-    commodity: str,
-    namespace: str,
-    value: str,
-    currency: str = "USD",
-    date: str | None = None,
-    price_type: str = "nav",
-    source: str = "user:price",
-) -> str:
-    """Record a price for a commodity (stock price, NAV, exchange rate).
-
-    Args:
-        commodity: Symbol of the commodity (e.g., "VTSAX", "AAPL").
-        namespace: Namespace of the commodity (e.g., "FUND", "NASDAQ").
-        value: Price per unit as decimal string (e.g., "250.45").
-        currency: Currency the price is denominated in. Default "USD".
-        date: Price date in ISO format (YYYY-MM-DD). Defaults to today.
-        price_type: Type of price:
-            - "nav" for mutual fund net asset value (default)
-            - "last" for last trade price
-            - "bid" / "ask" for bid/ask prices
-            - "unknown" for unspecified
-        source: Source identifier. Default "user:price".
-
-    Note:
-        If a price already exists for the same commodity/currency/date/source,
-        it will be updated rather than creating a duplicate.
-    """
-    book = get_book()
-    price_date = None
-    if date:
-        from datetime import date as date_type
-
-        price_date = date_type.fromisoformat(date)
-
-    result = book.create_price(
-        commodity=commodity,
-        namespace=namespace,
-        value=value,
-        currency=currency,
-        price_date=price_date,
-        price_type=price_type,
-        source=source,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_prices(
-    commodity: str,
-    namespace: str,
-    start_date: str | None = None,
-    end_date: str | None = None,
-    currency: str | None = None,
-) -> str:
-    """Get price history for a commodity.
-
-    Args:
-        commodity: Symbol of the commodity (e.g., "VTSAX").
-        namespace: Namespace of the commodity (e.g., "FUND").
-        start_date: Optional start date filter (YYYY-MM-DD).
-        end_date: Optional end date filter (YYYY-MM-DD).
-        currency: Optional currency filter (e.g., "USD").
-
-    Returns:
-        JSON with list of prices sorted by date descending (most recent first).
-    """
-    book = get_book()
-    from datetime import date as date_type
-
-    start = date_type.fromisoformat(start_date) if start_date else None
-    end = date_type.fromisoformat(end_date) if end_date else None
-
-    result = book.get_prices(
-        commodity=commodity,
-        namespace=namespace,
-        start_date=start,
-        end_date=end,
-        currency=currency,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_latest_price(
-    commodity: str,
-    namespace: str,
-    currency: str = "USD",
-) -> str:
-    """Get the most recent price for a commodity.
-
-    Args:
-        commodity: Symbol of the commodity (e.g., "VTSAX").
-        namespace: Namespace of the commodity (e.g., "FUND").
-        currency: Currency for the price. Default "USD".
-
-    Returns:
-        JSON with date, value, type, and source of most recent price.
-        Returns null if no price exists.
-    """
-    book = get_book()
-    result = book.get_latest_price(
-        commodity=commodity,
-        namespace=namespace,
-        currency=currency,
-    )
-    return _json(result)
+# Investments tools (commodities/prices/lots) moved to gnucash_mcp/tools/investments.py.
 
 
 @mcp.tool()
@@ -877,733 +708,7 @@ def replace_splits(
     return _json(result)
 
 
-# ============== Reconciliation Tools ==============
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="set_state", entity_type="split")
-def set_reconcile_state(
-    split_guid: Annotated[str, Field(description="GUID of the split to update (32-character hex string, or 8+ char prefix)")],
-    state: str,
-    reconcile_date: str | None = None,
-) -> str:
-    """Set the reconciliation state for a split.
-
-    Args:
-        split_guid: GUID of the split to update (32-character hex string, or 8+ char prefix)
-        state: New reconcile state: 'n' (new), 'c' (cleared), 'y' (reconciled)
-        reconcile_date: Date in ISO format (YYYY-MM-DD). Required for 'y', defaults to today.
-    """
-    book = get_book()
-    rec_date = date.fromisoformat(reconcile_date) if reconcile_date else None
-    result = book.set_reconcile_state(
-        split_guid=split_guid,
-        state=state,
-        reconcile_date=rec_date,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_unreconciled_splits(
-    account: str,
-    as_of_date: str | None = None,
-    verbose: bool = False,
-) -> str:
-    """Get all unreconciled splits for an account.
-
-    Returns a compact one-line-per-split format by default with a summary footer.
-    Use verbose=true for full JSON with split GUIDs, amounts, and totals.
-
-    Args:
-        account: Full account name (e.g., 'Assets:Bank:Checking')
-        as_of_date: Only include splits on or before this date (YYYY-MM-DD)
-        verbose: If true, return full JSON details. Default compact one-line format.
-    """
-    book = get_book()
-    date_obj = date.fromisoformat(as_of_date) if as_of_date else None
-    result = book.get_unreconciled_splits(
-        account_name=account,
-        as_of_date=date_obj,
-        compact=not verbose,
-    )
-    if verbose:
-        return _json(result)
-    return result
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="reconcile", entity_type="split")
-def reconcile_account(
-    account: str,
-    statement_date: str,
-    statement_balance: str,
-    split_guids: Annotated[list[str], Field(description="List of split GUIDs to mark as reconciled (8+ char prefixes accepted)")],
-) -> str:
-    """Reconcile multiple splits against a statement balance.
-
-    Marks all specified splits as reconciled if the resulting balance matches
-    the statement balance. This is an atomic operation - either all splits are
-    reconciled or none are.
-
-    Args:
-        account: Full account name (e.g., 'Assets:Bank:Checking')
-        statement_date: Statement ending date (YYYY-MM-DD)
-        statement_balance: Expected balance from statement (as string, e.g., '1234.56')
-        split_guids: List of split GUIDs to mark as reconciled (8+ char prefixes accepted)
-    """
-    book = get_book()
-    stmt_date = date.fromisoformat(statement_date)
-    result = book.reconcile_account(
-        account_name=account,
-        statement_date=stmt_date,
-        statement_balance=statement_balance,
-        split_guids=split_guids,
-    )
-    return _json(result)
-
-
-# ============== Void Tools ==============
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="void", entity_type="transaction")
-def void_transaction(
-    guid: Annotated[str, Field(description="Transaction GUID to void (32-character hex string, or 8+ char prefix)")],
-    reason: str,
-) -> str:
-    """Void a transaction (proper accounting void, not delete).
-
-    Voiding preserves the transaction for audit purposes but zeroes out
-    all split values. Use this instead of delete when you need to maintain
-    an audit trail.
-
-    Args:
-        guid: Transaction GUID to void (32-character hex string, or 8+ char prefix)
-        reason: Reason for voiding (required for audit trail)
-    """
-    book = get_book()
-    result = book.void_transaction(guid=guid, reason=reason)
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="unvoid", entity_type="transaction")
-def unvoid_transaction(
-    guid: Annotated[str, Field(description="Transaction GUID to unvoid (32-character hex string, or 8+ char prefix)")],
-) -> str:
-    """Restore a voided transaction.
-
-    Restores original split values and removes void markers.
-
-    Args:
-        guid: Transaction GUID to unvoid (32-character hex string, or 8+ char prefix)
-    """
-    book = get_book()
-    result = book.unvoid_transaction(guid=guid)
-    return _json(result)
-
-
-# ============== Reporting Tools ==============
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def spending_by_category(
-    start_date: str,
-    end_date: str,
-    depth: int = 1,
-) -> str:
-    """Get spending breakdown by expense category for a period.
-
-    Args:
-        start_date: Start of period (YYYY-MM-DD)
-        end_date: End of period (YYYY-MM-DD)
-        depth: Hierarchy depth for grouping (1 = top-level categories, 2 = subcategories)
-    """
-    book = get_book()
-    result = book.spending_by_category(
-        start_date=date.fromisoformat(start_date),
-        end_date=date.fromisoformat(end_date),
-        depth=depth,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def income_by_source(
-    start_date: str,
-    end_date: str,
-    depth: int = 1,
-) -> str:
-    """Get income breakdown by source for a period.
-
-    Args:
-        start_date: Start of period (YYYY-MM-DD)
-        end_date: End of period (YYYY-MM-DD)
-        depth: Hierarchy depth for grouping (1 = top-level categories, 2 = subcategories)
-    """
-    book = get_book()
-    result = book.income_by_source(
-        start_date=date.fromisoformat(start_date),
-        end_date=date.fromisoformat(end_date),
-        depth=depth,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def balance_sheet(as_of_date: str) -> str:
-    """Generate a balance sheet as of a specific date.
-
-    Shows assets, liabilities, and equity with account breakdowns.
-
-    Args:
-        as_of_date: Date to calculate balances as of (YYYY-MM-DD)
-    """
-    book = get_book()
-    result = book.balance_sheet(as_of_date=date.fromisoformat(as_of_date))
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def net_worth(
-    end_date: str,
-    start_date: str | None = None,
-    interval: str | None = None,
-) -> str:
-    """Calculate net worth (assets minus liabilities).
-
-    Can calculate a single point-in-time value or a time series.
-
-    Args:
-        end_date: Calculate net worth as of this date (YYYY-MM-DD)
-        start_date: Optional start date for time series (YYYY-MM-DD)
-        interval: Optional interval for time series: 'month', 'quarter', or 'year'
-    """
-    book = get_book()
-    result = book.net_worth(
-        end_date=date.fromisoformat(end_date),
-        start_date=date.fromisoformat(start_date) if start_date else None,
-        interval=interval,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def cash_flow(
-    start_date: str,
-    end_date: str,
-    account: str | None = None,
-) -> str:
-    """Calculate cash flow (inflows and outflows) for a period.
-
-    Args:
-        start_date: Start of period (YYYY-MM-DD)
-        end_date: End of period (YYYY-MM-DD)
-        account: Optional specific account to analyze (defaults to all cash/bank accounts)
-    """
-    book = get_book()
-    result = book.cash_flow(
-        start_date=date.fromisoformat(start_date),
-        end_date=date.fromisoformat(end_date),
-        account=account,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def debt_payoff_plan(
-    monthly_budget: str,
-    additional_purchase: str | None = None,
-) -> str:
-    """Calculate an avalanche-method debt payoff schedule with YETI multiplier.
-
-    Auto-discovers CREDIT/LIABILITY accounts that have an 'apr' slot set.
-    Set APRs via set_account_slot (e.g., set_account_slot("Liabilities:Visa", "apr", "23.49")).
-
-    YETI (Your Expense's True Impact) shows the true cost of a purchase when
-    carrying debt: "A $1.00 purchase will cost you $1.68 by the time your
-    debt is paid off."
-
-    Args:
-        monthly_budget: Total monthly amount available for all debt payments combined
-        additional_purchase: Dollar amount to calculate YETI for (default "1.00")
-    """
-    book = get_book()
-    result = book.debt_payoff_plan(
-        monthly_budget=monthly_budget,
-        additional_purchase=additional_purchase,
-    )
-    return _json(result)
-
-
-# ============== Budget Tools ==============
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def list_budgets() -> str:
-    """List all budgets in the book.
-
-    Returns:
-        JSON list of budgets with guid, name, description,
-        num_periods, and period_type.
-    """
-    book = get_book()
-    result = book.list_budgets()
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_budget(name: str) -> str:
-    """Get full details of a budget including all budget amounts.
-
-    Args:
-        name: Budget name.
-
-    Returns:
-        JSON with budget info and all account/period amounts.
-    """
-    book = get_book()
-    result = book.get_budget(name)
-    if result is None:
-        return _json({"error": f"Budget not found: {name}"})
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="create", entity_type="budget")
-def create_budget(
-    name: str,
-    year: int | None = None,
-    num_periods: int = 12,
-    period_type: str = "monthly",
-    description: str = "",
-) -> str:
-    """Create a new budget.
-
-    Args:
-        name: Budget name (e.g., "2026 Budget").
-        year: Budget year. Defaults to current year. Used to set start date.
-        num_periods: Number of periods. Default 12 (monthly for a year).
-        period_type: Period length:
-            - "monthly" (default)
-            - "quarterly"
-            - "weekly"
-        description: Optional description.
-    """
-    book = get_book()
-    result = book.create_budget(
-        name=name,
-        year=year,
-        num_periods=num_periods,
-        period_type=period_type,
-        description=description,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="update", entity_type="budget")
-def set_budget_amount(
-    budget_name: str,
-    account: str,
-    amount: str,
-    period: int | str | None = None,
-) -> str:
-    """Set a budget target for an account.
-
-    Args:
-        budget_name: Name of the budget.
-        account: Full account path (e.g., "Expenses:Groceries").
-        amount: Monthly budget amount as string (e.g., "500.00").
-        period: Which period(s) to set:
-            - None or "all": Set same amount for all periods (default)
-            - Integer 0-11: Set specific period (0 = January for yearly budget)
-            - "q1", "q2", "q3", "q4": Set all periods in quarter
-    """
-    book = get_book()
-    result = book.set_budget_amount(
-        budget_name=budget_name,
-        account=account,
-        amount=amount,
-        period=period,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_budget_report(
-    budget_name: str,
-    period: int | str | None = None,
-    account: str | None = None,
-    include_children: bool = True,
-) -> str:
-    """Compare actual spending against budget.
-
-    Args:
-        budget_name: Name of the budget.
-        period: Which period to report:
-            - None: Current period based on today's date (default)
-            - Integer 0-11: Specific period
-            - "ytd": Year to date (all periods up to current)
-            - "all": All periods
-        account: Optional filter to specific account or parent account.
-        include_children: If True and account specified, include child accounts.
-    """
-    book = get_book()
-    result = book.get_budget_report(
-        budget_name=budget_name,
-        period=period,
-        account=account,
-        include_children=include_children,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="delete", entity_type="budget")
-def delete_budget(name: str) -> str:
-    """Delete a budget.
-
-    Args:
-        name: Budget name.
-    """
-    book = get_book()
-    result = book.delete_budget(name=name)
-    return _json(result)
-
-
-# ============== Scheduled Transaction Tools ==============
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="create", entity_type="scheduled_transaction")
-def create_scheduled_transaction(
-    name: str,
-    description: str,
-    splits: list[dict],
-    start_date: str,
-    frequency: str,
-    end_date: str | None = None,
-    enabled: bool = True,
-) -> str:
-    """Create a recurring transaction template.
-
-    Args:
-        name: Name for the scheduled transaction (e.g., "Monthly Rent").
-        description: Transaction description when created.
-        splits: List of splits, same format as create_transaction:
-            [{"account": "Expenses:Rent", "amount": "1850.00"}, ...]
-        start_date: First occurrence date (YYYY-MM-DD).
-        frequency: How often it recurs:
-            - "weekly"
-            - "biweekly" (every 2 weeks)
-            - "monthly"
-            - "bimonthly" (every 2 months)
-            - "quarterly" (every 3 months)
-            - "yearly"
-        end_date: Optional last occurrence date (YYYY-MM-DD).
-        enabled: Whether the schedule is active. Default True.
-    """
-    book = get_book()
-    result = book.create_scheduled_transaction(
-        name=name,
-        description=description,
-        splits=splits,
-        start_date=start_date,
-        frequency=frequency,
-        end_date=end_date,
-        enabled=enabled,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def list_scheduled_transactions(
-    enabled_only: bool = True,
-    verbose: bool = False,
-) -> str:
-    """List all scheduled transactions.
-
-    Returns a compact one-line-per-schedule format by default.
-    Use verbose=true for full JSON with GUIDs, splits, dates, etc.
-
-    Args:
-        enabled_only: If True, only show enabled schedules. Default True.
-        verbose: If true, return full JSON details for each scheduled transaction.
-    """
-    book = get_book()
-    result = book.list_scheduled_transactions(
-        enabled_only=enabled_only,
-        compact=not verbose,
-    )
-    if verbose:
-        return _json(result)
-    return result
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_upcoming_transactions(
-    days: int = 14,
-    verbose: bool = False,
-) -> str:
-    """Get scheduled transactions due within a time window.
-
-    This is the "what bills are coming up?" query.
-
-    Args:
-        days: Look ahead window in days. Default 14.
-        verbose: If true, return full JSON with splits. Default compact one-line format.
-    """
-    book = get_book()
-    result = book.get_upcoming_transactions(days=days, compact=not verbose)
-    if verbose:
-        return _json(result)
-    return result
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="create", entity_type="transaction")
-def create_transaction_from_scheduled(
-    guid: Annotated[str, Field(description="Scheduled transaction GUID (or 8+ char prefix)")],
-    transaction_date: str | None = None,
-) -> str:
-    """Create an actual transaction from a scheduled template.
-
-    Args:
-        guid: Scheduled transaction GUID (or 8+ char prefix).
-        transaction_date: Date for the transaction. Defaults to next occurrence.
-    """
-    book = get_book()
-    result = book.create_transaction_from_scheduled(
-        guid=guid,
-        transaction_date=transaction_date,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="update", entity_type="scheduled_transaction")
-def update_scheduled_transaction(
-    guid: Annotated[str, Field(description="Scheduled transaction GUID (or 8+ char prefix)")],
-    enabled: bool | None = None,
-    end_date: str | None = None,
-) -> str:
-    """Update a scheduled transaction.
-
-    Args:
-        guid: Scheduled transaction GUID (or 8+ char prefix).
-        enabled: Enable or disable.
-        end_date: Set end date (empty string to clear).
-    """
-    book = get_book()
-    result = book.update_scheduled_transaction(
-        guid=guid,
-        enabled=enabled,
-        end_date=end_date,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="delete", entity_type="scheduled_transaction")
-def delete_scheduled_transaction(
-    guid: Annotated[str, Field(description="Scheduled transaction GUID (or 8+ char prefix)")],
-) -> str:
-    """Delete a scheduled transaction.
-
-    Does not affect transactions already created from this schedule.
-
-    Args:
-        guid: Scheduled transaction GUID (or 8+ char prefix).
-    """
-    book = get_book()
-    result = book.delete_scheduled_transaction(guid=guid)
-    return _json(result)
-
-
-# ============== Lot (Cost Basis) Tools ==============
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="create", entity_type="lot")
-def create_lot(
-    account: str,
-    title: str,
-    notes: str = "",
-) -> str:
-    """Create a new lot for cost basis tracking.
-
-    Lots group investment purchases for tracking cost basis and
-    calculating capital gains when selling.
-
-    Args:
-        account: Full path of investment account (e.g., "Assets:Investments:VTSAX").
-        title: Lot identifier (e.g., "VTSAX 2026-01-15 purchase").
-        notes: Optional notes.
-    """
-    book = get_book()
-    result = book.create_lot(account=account, title=title, notes=notes)
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def list_lots(
-    account: str,
-    include_closed: bool = False,
-    verbose: bool = False,
-) -> str:
-    """List all lots for an investment account.
-
-    Returns a compact one-line-per-lot format by default.
-    Use verbose=true for full JSON with guid, title, notes, etc.
-
-    Args:
-        account: Full path of investment account.
-        include_closed: If True, include fully-sold lots. Default False.
-        verbose: If true, return full JSON details for each lot.
-    """
-    book = get_book()
-    result = book.list_lots(account=account, include_closed=include_closed, compact=not verbose)
-    if verbose:
-        return _json(result)
-    return result
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def get_lot(
-    guid: Annotated[str, Field(description="Lot GUID (or 8+ char prefix)")],
-) -> str:
-    """Get detailed information about a lot.
-
-    Args:
-        guid: Lot GUID (or 8+ char prefix).
-
-    Returns:
-        JSON with lot details including all splits:
-        - title, notes, is_closed
-        - splits: list of all splits with date, quantity, value
-        - summary: total quantity, cost basis, cost per share
-    """
-    book = get_book()
-    result = book.get_lot(guid=guid)
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="update", entity_type="lot")
-def assign_split_to_lot(
-    split_guid: Annotated[str, Field(description="GUID of the split (from transaction's investment account). 8+ char prefix accepted.")],
-    lot_guid: Annotated[str, Field(description="GUID of the lot (or 8+ char prefix)")],
-) -> str:
-    """Assign a transaction split to a lot.
-
-    Use after creating a buy/sell transaction to link the investment
-    account split to its lot for cost basis tracking.
-
-    Args:
-        split_guid: GUID of the split (from transaction's investment account). 8+ char prefix accepted.
-        lot_guid: GUID of the lot (or 8+ char prefix).
-
-    Workflow:
-        1. create_lot("Assets:VTSAX", "VTSAX Jan 2026")
-        2. create_transaction(...buy 10 shares...)
-        3. assign_split_to_lot(investment_split_guid, lot_guid)
-    """
-    book = get_book()
-    result = book.assign_split_to_lot(split_guid=split_guid, lot_guid=lot_guid)
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="read")
-def calculate_lot_gain(
-    lot_guid: Annotated[str, Field(description="Lot GUID (or 8+ char prefix)")],
-    shares: str | None = None,
-    sale_price: str | None = None,
-) -> str:
-    """Calculate potential or actual capital gain for a lot.
-
-    If shares and sale_price provided, calculates hypothetical gain.
-    Otherwise uses lot's current state and latest price.
-
-    Args:
-        lot_guid: Lot GUID (or 8+ char prefix).
-        shares: Optional number of shares to calculate for.
-                Defaults to all remaining shares.
-        sale_price: Optional sale price per share.
-                    Defaults to latest price for the commodity.
-    """
-    book = get_book()
-    result = book.calculate_lot_gain(
-        lot_guid=lot_guid, shares=shares, sale_price=sale_price,
-    )
-    return _json(result)
-
-
-@mcp.tool()
-@safe_tool
-@audit_log(classification="write", operation="update", entity_type="lot")
-def close_lot(
-    guid: Annotated[str, Field(description="Lot GUID (or 8+ char prefix)")],
-) -> str:
-    """Mark a lot as closed.
-
-    Use when a lot is fully sold but wasn't automatically marked closed,
-    or to manually close a lot with zero shares.
-
-    Args:
-        guid: Lot GUID (or 8+ char prefix).
-
-    Note:
-        Lots are automatically marked closed when their quantity reaches zero
-        through assigned splits. This tool is for manual cleanup.
-    """
-    book = get_book()
-    result = book.close_lot(guid=guid)
-    return _json(result)
+# Reconciliation/reporting/budgets/scheduling/lots tools moved to gnucash_mcp/tools/<module>.py.
 
 
 # Admin tools (get_account_slots, set_account_slot, delete_account_slot,
